requirements.txt
django>=4.2.2
python-dotenv>=1.0.0
mysqlclient>=2.2.0
djangorestframework>=3.14.0
django-filter>=23.2
drf-spectacular>=0.26.3
